translation/backend/requirements.txt
# ========================================================
# Python Dependencies for Multilingual Translation System
# ========================================================

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.9

# Machine Learning & NLP
torch>=2.6.0
transformers>=4.40.0
datasets>=2.18.0
sentencepiece>=0.2.0
protobuf>=4.25.0

# Translation & Evaluation
sacrebleu>=2.4.0
evaluate>=0.4.1

# Speech Processing
openai-whisper>=20231117
gtts>=2.5.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Utilities
tqdm>=4.66.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0
